src/tensorrt/tensorrt-5.0.2.6/samples-python/python/uff_ssd/requirements.txt
numpy
Pillow
pycuda
requests
tensorflow-gpu